pyproject.toml
[project]
name = "inegi-mcp"
version = "0.1.0"
description = "MCP Server para acceder a las APIs del INEGI (Indicadores y DENUE)"
authors = [
    {name = "Guillermo Chávez", email = "guillermo830903@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
inegi-mcp = "inegi_mcp.server:main"
